+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # CHANGELOG
 
+15 March 2025
+- You can now pause and resume the timer
+
 14 March 2025
 - Initial commit